Fix playing glitch
When a note was repeated (say, two eighth notes of C₅ immediately
following one another), there would be two events in the queue to be
executed at the same time: Note1_off and Note2_on. If those are executed
in the wrong order, Note1_off prematurely ends Note2 immediately after
it is started.

Since "noteoff" < "noteon", we simply compare the strings as part of the
ordering process, and `heapq` takes care of the rest for us.